requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.9.24
click==8.1.3
colorama==0.4.5
distlib==0.3.6
Django==4.1.1
django-extensions==3.2.1
djangorestframework==3.14.0
filelock==3.8.0
flake8==5.0.4
graphviz==0.20.1
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
pipenv==2022.9.24
platformdirs==2.5.2
pycodestyle==2.9.1
pydot==1.4.2
pyflakes==2.5.0
pyparsing==3.0.9
pytz==2022.4
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
virtualenv==20.16.5
virtualenv-clone==0.5.7
